Add --custom option to customize kernel

+++ build.sh
@@ -5,6 +5,7 @@ export LANG=C
 update=yes
 btype=binary
 shell=no
+custom=no
 
 args=( $@ );
 for (( i=0; $i < $# ; i++ ))
@@ -12,6 +13,7 @@ do
   [[ "${args[$i]}" =~ --update.* ]] && update=${args[$i]#*=} && continue
   [[ "${args[$i]}" =~ --btype.* ]] && btype=${args[$i]#*=} && continue
   [[ "${args[$i]}" =~ --shell.* ]] && shell=${args[$i]#*=} && continue
+  [[ "${args[$i]}" =~ --custom.* ]] && custom=${args[$i]#*=} && continue
 done
 
 cd $ksrc
@@ -22,28 +24,42 @@ then
   bash
 fi
 
+echo -e "********\n\nCleaning git source tree\n\n********"
 git clean -fdx
 git reset --hard HEAD
 
 if [ "$update" == "yes" ]
 then
+  echo -e "********\n\nUpdating git source tree\n\n********"
   git checkout master
   git fetch origin 
   git pull
 fi
 
 # checkout the kver
+echo -e "********\n\nSwitching to cod/mainline/${kver} branch\n\n********"
 git checkout "cod/mainline/${kver}"
 
 # prep
+echo -e "********\n\nApplying default configs\n\n********"
 fakeroot debian/rules clean defaultconfigs
 fakeroot debian/rules clean
 
+if [ "$custom" == "yes" ]
+then
+  echo -e "********\n\nYou have asked for a custom build.\nYou will be given the chance to makemenuconfig next.\n\n********"
+  read -p 'Press return to continue...' foo
+  fakeroot debian/rules editconfigs
+fi
+
 # Build
+echo -e "********\n\nBuilding packages\n\n********"
 dpkg-buildpackage -uc -ui -us -aamd64 -d --build=$btype
 #AUTOBUILD=1 fakeroot debian/rules binary-debs
 #AUTOBUILD=1 NOEXTRAS=1 fakeroot debian/rules binary-FLAVOUR
 
+echo -e "********\n\nMoving packages to debs folder\n\n********"
+dpkg-buildpackage -uc -ui -us -aamd64 -d --build=$btype
 mv $ksrc/../*.* $kdeb
 
 if [ "$shell" == "yes" ]
@@ -52,4 +68,8 @@ then
   bash
 fi
 
+echo -e "********\n\nCleaning git source tree\n\n********"
+git clean -fdx
+git reset --hard HEAD
+git checkout master
